storage layer: standardize on (uint8_t whichkey)
Before this commit, the storage layer (storage.h) had 3 functions
which accepted (uint8_t whichkey):
- storage_transaction_checkpoint()
- storage_transaction_commit()
- storage_transaction_commitfromcheckpoint

There was 1 function which accepted (int whichkey):
- storage_fsck_start()

Changing storage_fsck_start() to take an uint_8 sets off a chain:
    storage.h			storage_fsck_start()
    storage_delete.c		storage_fsck_start()
whichkey is passed to:
    storage_internal.h		storage_transaction_start_fsck()
    storage_transaction.c	storage_transaction_start_fsck()
which merely tests
    if (whichkey)
        ...
    else
        ...
so it doesn't matter if this is an int or not.

+++ tar/storage/storage.h
@@ -139,7 +139,7 @@ STORAGE_D * storage_delete_start(uint64_t, const uint8_t[32], uint8_t[32]);
  * transaction into ${seqnum}.  If ${whichkey} is zero, use the write key
  * (in which case the transaction must be readonly).
  */
-STORAGE_D * storage_fsck_start(uint64_t, uint8_t[32], int, int);
+STORAGE_D * storage_fsck_start(uint64_t, uint8_t[32], int, uint8_t);
 
 /**
  * storage_delete_file(S, class, name):
